Renaming the file for namespaces.
