+++ src/namespaces.ts
